+++ src/dotenvx/__init__.py
@@ -1,1 +1,3 @@
-from .main import load_dotenvx
+from .main import dotenv_values, find_dotenv, load_dotenv, load_dotenvx
+
+__all__ = ["dotenv_values", "find_dotenv", "load_dotenv", "load_dotenvx"]